Refactor directory structure for better organization
