+++ foundrytools_cli_2/cli/cff/cli.py
@@ -13,7 +13,7 @@ from foundrytools_cli_2.cli.cff.options import (
 )
 from foundrytools_cli_2.cli.shared_callbacks import validate_params
 from foundrytools_cli_2.cli.shared_options import base_options, new_string, old_string
-from foundrytools_cli_2.cli.task_runner import TaskRunner
+from foundrytools_cli_2.lib.cli_tools.task_runner import TaskRunner
 
 cli = click.Group("cff", help="Utilities for editing the ``CFF`` table.")
