+++ foundrytools_cli_2/cli/converter/tasks/ttf_to_otf.py
@@ -5,9 +5,9 @@ from afdko.fdkutils import run_shell_command
 
 from foundrytools_cli_2.cli.logger import logger
 from foundrytools_cli_2.lib.font import Font
-from foundrytools_cli_2.lib.font.tables import OS2Table
-from foundrytools_cli_2.lib.otf.otf_builder import build_otf
-from foundrytools_cli_2.lib.otf.t2_charstrings import quadratics_to_cubics_2
+from foundrytools_cli_2.lib.otf_builder import build_otf
+from foundrytools_cli_2.lib.t2_charstrings import quadratics_to_cubics_2
+from foundrytools_cli_2.lib.tables import OS2Table
 
 
 def _build_out_file_name(font: Font, output_dir: t.Optional[Path], overwrite: bool = True) -> Path:
