+++ foundrytools_cli_2/cli/gsub/cli.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import click
 
 from foundrytools_cli_2.cli.shared_options import base_options
-from foundrytools_cli_2.cli.task_runner import TaskRunner
+from foundrytools_cli_2.lib.cli_tools.task_runner import TaskRunner
 
 cli = click.Group(help="Utilities for editing the ``GSUB`` table.")
