+++ foundrytools_cli_2/cli/otf/cli.py
@@ -6,7 +6,7 @@ import click
 
 from foundrytools_cli_2.cli.otf.options import drop_hinting_data_flag, otf_autohint_options
 from foundrytools_cli_2.cli.shared_options import base_options, subroutinize_flag
-from foundrytools_cli_2.cli.task_runner import TaskRunner
+from foundrytools_cli_2.lib.cli_tools.task_runner import TaskRunner
 from foundrytools_cli_2.lib.font import Font
 
 cli = click.Group(help="Utilities for editing OpenType-PS fonts.")
