+++ foundrytools_cli_2/cli/post/tasks.py
@@ -2,7 +2,7 @@ import typing as t
 
 from foundrytools_cli_2.cli.logger import logger
 from foundrytools_cli_2.lib.font import Font
-from foundrytools_cli_2.lib.font.tables import PostTable
+from foundrytools_cli_2.lib.tables import PostTable
 
 
 def set_attrs(
