+++ foundrytools_cli_2/cli/utils/cli.py
@@ -5,8 +5,8 @@ from pathlib import Path
 import click
 
 from foundrytools_cli_2.cli.shared_options import input_path_argument, recursive_flag
-from foundrytools_cli_2.cli.task_runner import TaskRunner
 from foundrytools_cli_2.cli.utils.options import font_organizer_options, rename_source_option
+from foundrytools_cli_2.lib.cli_tools.task_runner import TaskRunner
 
 cli = click.Group(help="Miscellaneous utilities.")
