--- foundrytools_cli_2/lib/beziers/__init__.py
@@ -1,3 +1,0 @@
-from foundrytools_cli_2.lib.beziers.beziers_tools import add_extremes
-
-__all__ = ["add_extremes"]
