--- foundrytools_cli_2/lib/font/__init__.py
@@ -1,3 +1,0 @@
-from foundrytools_cli_2.lib.font.font import Font
-
-__all__ = ["Font"]
