--- foundrytools_cli_2/lib/font/tables/__init__.py
@@ -1,23 +1,0 @@
-from foundrytools_cli_2.lib.font.tables.cff_ import CFFTable
-from foundrytools_cli_2.lib.font.tables.gdef import GdefTable
-from foundrytools_cli_2.lib.font.tables.glyf import GlyfTable
-from foundrytools_cli_2.lib.font.tables.gsub import GsubTable
-from foundrytools_cli_2.lib.font.tables.head import HeadTable
-from foundrytools_cli_2.lib.font.tables.hhea import HheaTable
-from foundrytools_cli_2.lib.font.tables.hmtx import HmtxTable
-from foundrytools_cli_2.lib.font.tables.name import NameTable
-from foundrytools_cli_2.lib.font.tables.os_2 import OS2Table
-from foundrytools_cli_2.lib.font.tables.post import PostTable
-
-__all__ = [
-    "CFFTable",
-    "GdefTable",
-    "GlyfTable",
-    "GsubTable",
-    "HeadTable",
-    "HheaTable",
-    "HmtxTable",
-    "NameTable",
-    "OS2Table",
-    "PostTable",
-]
