+++ foundrytools_cli_2/lib/tables/__init__.py
@@ -1,0 +1,25 @@
+from foundrytools_cli_2.lib.tables.cff_ import CFFTable
+from foundrytools_cli_2.lib.tables.cmap import CmapTable
+from foundrytools_cli_2.lib.tables.gdef import GdefTable
+from foundrytools_cli_2.lib.tables.glyf import GlyfTable
+from foundrytools_cli_2.lib.tables.gsub import GsubTable
+from foundrytools_cli_2.lib.tables.head import HeadTable
+from foundrytools_cli_2.lib.tables.hhea import HheaTable
+from foundrytools_cli_2.lib.tables.hmtx import HmtxTable
+from foundrytools_cli_2.lib.tables.name import NameTable
+from foundrytools_cli_2.lib.tables.os_2 import OS2Table
+from foundrytools_cli_2.lib.tables.post import PostTable
+
+__all__ = [
+    "CFFTable",
+    "CmapTable",
+    "GdefTable",
+    "GlyfTable",
+    "GsubTable",
+    "HeadTable",
+    "HheaTable",
+    "HmtxTable",
+    "NameTable",
+    "OS2Table",
+    "PostTable",
+]
