+++ foundrytools_cli_2/lib/tables/cmap.py
@@ -3,7 +3,7 @@ import typing as t
 from fontTools.ttLib import TTFont
 
 from foundrytools_cli_2.lib.constants import T_CMAP
-from foundrytools_cli_2.lib.font.tables.default import DefaultTbl
+from foundrytools_cli_2.lib.tables.default import DefaultTbl
 
 
 class CmapTable(DefaultTbl):  # pylint: disable=too-few-public-methods
