+++ foundrytools_cli_2/lib/tables/head.py
@@ -2,7 +2,7 @@ from fontTools.misc.textTools import num2binary
 from fontTools.ttLib import TTFont
 
 from foundrytools_cli_2.lib.constants import T_HEAD
-from foundrytools_cli_2.lib.font.tables.default import DefaultTbl
+from foundrytools_cli_2.lib.tables.default import DefaultTbl
 from foundrytools_cli_2.lib.utils.bits_tools import is_nth_bit_set
 
 BOLD_BIT = 0
