+++ foundrytools_cli_2/lib/tables/hmtx.py
@@ -1,7 +1,7 @@
 from fontTools.ttLib import TTFont
 
 from foundrytools_cli_2.lib.constants import T_HMTX
-from foundrytools_cli_2.lib.font.tables.default import DefaultTbl
+from foundrytools_cli_2.lib.tables.default import DefaultTbl
 
 
 class HmtxTable(DefaultTbl):  # pylint: disable=too-few-public-methods
