+++ foundrytools_cli_2/lib/tables/post.py
@@ -1,7 +1,7 @@
 from fontTools.ttLib import TTFont
 
 from foundrytools_cli_2.lib.constants import T_POST
-from foundrytools_cli_2.lib.font.tables.default import DefaultTbl
+from foundrytools_cli_2.lib.tables.default import DefaultTbl
 
 
 class PostTable(DefaultTbl):
